UI2/requirements.txt
Flask==3.1.1
Jinja2==3.1.6
Werkzeug==3.1.3
itsdangerous==2.2.0
SQLAlchemy==2.0.41
requests==2.32.4
python-dotenv==1.1.0
openai==1.107.1
tiktoken==0.9.0
faiss-cpu==1.11.0
numpy==1.26.4
pandas==2.3.0
beautifulsoup4==4.13.4
reportlab==3.6.8
mistral==20.0.0
mistral-lib==3.3.1
mistralai==1.8.2
openai==1.107.1
opentrons==8.4.1
opentrons_shared_data==8.4.1
fastjsonschema==2.21.2
jsonschema==4.17.3